+++ components/include/memfault/core/arch.h
@@ -17,6 +17,9 @@ extern "C" {
 //! Return true if the code is currently running in an interrupt context, false otherwise
 bool memfault_arch_is_inside_isr(void);
 
+//! Disable any configurable fault handlers supported by the platform
+void memfault_arch_disable_configurable_faults(void);
+
 #ifdef __cplusplus
 }
 #endif