+++ ports/zephyr/common/memfault_platform_debug_log.c
@@ -8,45 +8,59 @@
 
 #include "memfault/core/platform/debug_log.h"
 
+//
+
+#include <logging/log.h>
 #include <stdio.h>
 
+#include "memfault/config.h"
+#include "memfault/ports/zephyr/version.h"
 #include "zephyr_release_specific_headers.h"
 
-#include "memfault/config.h"
+LOG_MODULE_REGISTER(mflt, CONFIG_MEMFAULT_LOG_LEVEL);
 
 #ifndef MEMFAULT_DEBUG_LOG_BUFFER_SIZE_BYTES
-#  define MEMFAULT_DEBUG_LOG_BUFFER_SIZE_BYTES (128)
+  #define MEMFAULT_DEBUG_LOG_BUFFER_SIZE_BYTES (128)
 #endif
 
+//! Translate Memfault logs to Zephyr logs.
 void memfault_platform_log(eMemfaultPlatformLogLevel level, const char *fmt, ...) {
   va_list args;
   va_start(args, fmt);
 
   char log_buf[MEMFAULT_DEBUG_LOG_BUFFER_SIZE_BYTES];
   vsnprintf(log_buf, sizeof(log_buf), fmt, args);
+  char *log_str = log_buf;
 
-  const char *lvl_str = "???";
+#if !MEMFAULT_ZEPHYR_VERSION_GT(3, 1)
+  #if !defined(CONFIG_LOG2)
+  // Before zephyr 3.1, LOG was a different option from LOG2 and required
+  // manually duplicating string argument values.
+  log_str = log_strdup(log_buf);
+  #endif
+#endif
+
   switch (level) {
     case kMemfaultPlatformLogLevel_Debug:
-      lvl_str = "dbg";
+      LOG_DBG("%s", log_str);
       break;
 
     case kMemfaultPlatformLogLevel_Info:
-      lvl_str = "inf";
+      LOG_INF("%s", log_str);
       break;
 
     case kMemfaultPlatformLogLevel_Warning:
-      lvl_str = "wrn";
+      LOG_WRN("%s", log_str);
       break;
 
     case kMemfaultPlatformLogLevel_Error:
-      lvl_str = "err";
+      LOG_ERR("%s", log_str);
       break;
 
     default:
+      LOG_ERR("??? %s", log_str);
       break;
   }
-  printk("<%s> <mflt>: %s\n", lvl_str, log_buf);
 
   va_end(args);
 }
@@ -54,7 +68,18 @@ void memfault_platform_log(eMemfaultPlatformLogLevel level, const char *fmt, ...
 void memfault_platform_log_raw(const char *fmt, ...) {
   va_list args;
   va_start(args, fmt);
-  vprintk(fmt, args);
+
+#define ZEPHYR_VERSION_GTE(major, minor) \
+  ((KERNEL_VERSION_MAJOR > (major)) ||   \
+   ((KERNEL_VERSION_MAJOR == (major)) && (KERNEL_VERSION_MINOR >= (minor))))
+
+#if ZEPHYR_VERSION_GTE(3, 0)
+  char log_buf[MEMFAULT_DEBUG_LOG_BUFFER_SIZE_BYTES];
+  vsnprintf(log_buf, sizeof(log_buf), fmt, args);
+  LOG_PRINTK("%s", log_buf);
+#else
+  log_printk("%s\n", args);
+#endif
+
   va_end(args);
-  printk("\n");
 }